% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcdWrapper.R
\name{bcdFit}
\alias{bcdFit}
\title{Fit model by block coordinate descent}
\usage{
bcdFit(X, y, k = 1, family = c("gaussian", "multinomial", "poisson",
  "logistic"), groups, penaltyFactor, lambda = NULL,
  lambdaMinRatio = 1e-04, nLambda = 100, sampleWeights = NULL,
  maxit = 1e+05, tolerance = 10^-12, offset = NULL,
  eigenValueTolerance = 10^-9, rescale = TRUE, useDevTol = TRUE,
  devTol = 0.9)
}
\arguments{
\item{X}{design matrix}

\item{y}{response vector (or matrix)}

\item{k}{number of response categories (multinomial) or dimension of response matrix (Gaussian)}

\item{family}{type of regression: one of "gaussian", "multinomial", "poisson", or "logistic"}

\item{groups}{list containing the group index of each column}

\item{penaltyFactor}{vector containing penalty level for each group}

\item{lambda}{penalty vector for model fitting}

\item{lambdaMinRatio}{value to use for lambdaMin/lambdaMax (defaults to 0.0001)}

\item{nLambda}{number of lambda values to use}

\item{sampleWeights}{different weights for each sample}

\item{maxit}{maximum number of coordinate descent iterations (default 100000)}

\item{tolerance}{solution tolerance (default 10^-12)}

\item{offset}{offset to use for Poisson regression}

\item{eigenValueTolerance}{tolerance for deciding whether columns in a group are linearly independent (default 10^-9)}

\item{rescale}{force columns to have norm 1 after orthogonalization (defaults to TRUE)}

\item{useDevTol}{use deviance tolerance}

\item{devTol}{tolerance for deviance}
}
\description{
Fits a model by block coordinate descent
}
\examples{
#no example

}
