% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{beta}
\alias{beta}
\title{True coefficients for tests}
\format{Length 50 numeric vector}
\usage{
beta
}
\description{
Length 50 coefficient vector for all tests except
multinomial and multiresponse Gaussian. The 11th 
through 30th coefficients are all 0. The remainder
are nonzero.
}
\keyword{datasets}
