% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{beta_multiresponse}
\alias{beta_multiresponse}
\title{True coefficients for multiresponse tests}
\format{50 by 3 numeric matrix}
\usage{
beta_multiresponse
}
\description{
50 by 3 coefficient matrix in the tests for
multiresponse Gaussian and multinomial tests. 
The coefficients in rows 11-30 are all 0.
}
\keyword{datasets}
