\name{combine_labels}
\alias{combine_labels}
\title{Combine labels for different datasets}
\usage{
combine_labels(labels, keep_sample_name = TRUE, data = NULL, data_width = NULL)
}
\description{
Combine labels for different datasets.
}
\examples{
combine_labels(labels, keep_sample_name = TRUE, data = NULL, data_width = NULL)
}
