% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbmgs_mod.R
\name{fbmgs_mod}
\alias{fbmgs_mod}
\title{Discovery of Linear Structural Equation Models}
\usage{
fbmgs_mod(Y, moments = 4, meanEst = F, outerTol = 0.001,
  innerTol = 0.001, printStatus = F, fwdNullTol = 30, fwdDirectTol = 15,
  bwdNullTol = 15, bwdDirectTol = 15)
}
\arguments{
\item{Y}{data in V x N matrix}

\item{moments}{number of moments to test in estimating equations}

\item{meanEst}{whether or not to fit a mean}

\item{innerTol}{tolerance for dual problem}

\item{printStatus}{whether or not to print output}

\item{fwdTol}{tolerance for difference on forward search}

\item{bwdTol}{tolerance for difference on backward search}

\item{fwdType}{type of fit for forward search: profile or euclid}

\item{bwdType}{type of fit for backward search: profile or euclid}

\item{optimTol}{tolerance for optim}
}
\description{
Produces estimated SEM from edge-wise tests
}

