% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSemEl.R
\name{fitEL}
\alias{fitEL}
\title{Fitting Linear Structural Equation Models}
\usage{
fitEL(Y, B, Omega, B.hat = NULL, Omega.hat = NULL, mu.hat = NULL,
  optim.method = "BFGS", tol = 1e-06, maxInnerIter = 500,
  maxOuterIter = 100, meanEst = 0, type = "profile", high_moments = 2)
}
\arguments{
\item{B}{V by V matrix with {0,1} giving structure of directed edges}

\item{Omega}{V by V matrix with {0,1} giving structure of bi-directed edges}

\item{B.hat}{V by V matrix giving initial edges weights for directed edges. If BInit is NULL,
a default initialization will be used.}

\item{Omega.hat}{V by V matrix giving initial edge weights for bi-directed edges}

\item{mu.hat}{Initial value for means}

\item{type}{profile, naive, euclid}
}
\description{
Fits Empirical Likelihood estimates for given SEM
}

