% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forwardBackward_mod.R
\name{forwardBackwardMod}
\alias{forwardBackwardMod}
\title{Discovery of Linear Structural Equation Models}
\usage{
forwardBackwardMod(Y, moments = 3, meanEst = 1, optimTol = 1e-04,
  innerTol = 1e-04, printStatus = T, exch = F, vsNull = T,
  fwdTol = 30, fwdType = "profile", bwdTol = 15, bwdType = "euclid")
}
\arguments{
\item{Y}{data in V x N matrix}

\item{moments}{number of moments to test in estimating equations}

\item{meanEst}{whether or not to fit a mean}

\item{optimTol}{tolerance for optim}

\item{innerTol}{tolerance for dual problem}

\item{printStatus}{whether or not to print output}

\item{fwdTol}{tolerance for difference on forward search}

\item{fwdType}{type of fit for forward search: profile or euclid}

\item{bwdTol}{tolerance for difference on backward search}

\item{bwdType}{type of fit for backward search: profile or euclid}
}
\description{
Produces estimated SEM from edge-wise tests
}

