% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateModel.R
\name{generateModel}
\alias{generateModel}
\title{Generate Data from Structural Equation Model}
\usage{
generateModel(v, n, k, d, b, errorDist = "gauss")
}
\arguments{
\item{v}{number of nodes}

\item{n}{number of samples}

\item{k}{the size of the largest cycle}

\item{d}{the}
}
\value{
\item{sigmaHat}{estimated covariance matrix at convergence}
   \item{bHat}{estimated B matrix (edge weights for directed edges) at converegence}
   \item{omegaHat}{estimated Omega (edge weights for bi-directed edges) at convergence}
   \item{iterations}{number of iterations until convergence. a single iteration is considered
   a pass through all nodes}
   \item{converged}{boolean on whether or not the algorithm converged before the max iterations}
}
\description{
Samples a random Graph and Data
}
