% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_CI.R
\name{getELVariance}
\alias{getELVariance}
\title{Empirical Likelihood Methods for Linear Structural Equation Models}
\usage{
getELVariance(Y, B, Omega, B.hat, Omega.hat, p)
}
\arguments{
\item{Y}{V by N matrix with observed data}

\item{B}{V by V matrix with {0,1} giving structure of directed edges}

\item{Omega}{V by V matrix with {0,1} giving structure of bi-directed edges}

\item{B.hat}{V by V matrix giving estimated edges weights for directed edges}

\item{Omega.hat}{V by V matrix giving estimated edge weights for bi-directed edges}

\item{p}{vector of length N containing the estimated empirical likelihood weights}
}
\value{
The asymptotic covariance matrix as derived in Qin and Lawless 1994
}
\description{
Calculate assymptotic variance for non-structural zeros in B and Omega matrix. The ordering of the estimates
in the variance matrix follows standard vec(-) convention indexing over rows first then columns. The elements of B
are ordered before the elements of Omega. Duplicate elements of Omega are not included.
}

