% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_CI.R
\name{getRicfSandwich}
\alias{getRicfSandwich}
\title{Maximum Likelihood Method for Linear Structural Equation Models}
\usage{
getRicfSandwich(Y, B, Omega, B.hat, Omega.hat)
}
\arguments{
\item{Y}{V by N matrix corresponding to the observed sample}

\item{B}{V by V matrix with {0,1} giving structure of directed edges}

\item{Omega}{V by V matrix with {0,1} giving structure of bi-directed edges}

\item{B.hat}{V by V matrix giving estimated edges weights for directed edges}

\item{Omega.hat}{V by V matrix giving estimated edge weights for bi-directed edges}
}
\value{
The Huber-White Misspecified model covariance estimate
}
\description{
Calculate sandwich standard errors (for misspecified model) information for non-structural zeros in B and Omega matrix. The ordering of the estimates
in the Information matrix follows standard vec(-) convention indexing over rows first then columns. The elements of B
are ordered before the elements of Omega. Duplicate elements of Omega are not included.
}

