% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{grouping}
\alias{grouping}
\title{Grouping used for the example tests}
\format{
Length 4 list of integer vectors.
}
\usage{
grouping
}
\description{
Grouping used for the examples. Length 4 list of 
integer vectors.
}
\keyword{datasets}
