% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{penaltyFactor}
\alias{penaltyFactor}
\title{penaltyFactor}
\format{
Length 4 numeric vector
}
\usage{
penaltyFactor
}
\description{
Length 4 vector containing penalization level for groups
in the examples: c(0,1,1,1). Intercept group unpenalized.
}
\keyword{datasets}
