% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{reference_Linear}
\alias{reference_Linear}
\title{Linear regression reference coefficients}
\format{
Length 100 list of 50 by 1 numeric matrices
}
\usage{
reference_Linear
}
\description{
Linear regression reference coefficients for test without sample weights.
}
\keyword{datasets}
