% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{reference_Linear_Weighted}
\alias{reference_Linear_Weighted}
\title{Linear regression reference coefficients}
\format{Length 100 list of 50 by 1 numeric matrices}
\usage{
reference_Linear_Weighted
}
\description{
Linear regression reference coefficients for test with sample weights.
}
\keyword{datasets}
