% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{reference_Logistic}
\alias{reference_Logistic}
\title{Logistic regression reference coefficients}
\format{Length 100 list of 50 by 2 numeric matrices}
\usage{
reference_Logistic
}
\description{
Logistic regression reference coefficients for test without sample weights.
}
\keyword{datasets}
