% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{reference_Multinomial_Weighted}
\alias{reference_Multinomial_Weighted}
\title{Multinomial regression reference coefficients}
\format{
Length 100 list of 50 by 3 numeric matrices
}
\usage{
reference_Multinomial_Weighted
}
\description{
Multinomial regression reference coefficients for test with sample weights.
}
\keyword{datasets}
