% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{reference_Multiresponse}
\alias{reference_Multiresponse}
\title{Multiresponse Gaussian reference coefficients}
\format{Length 100 list of 50 by 3 numeric matrices}
\usage{
reference_Multiresponse
}
\description{
Multiresponse Gaussian reference coefficients for test without sample weights.
}
\keyword{datasets}
