% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{reference_Poisson}
\alias{reference_Poisson}
\title{Poisson regression reference coefficients}
\format{
Length 100 list of 50 by 1 numeric matrices
}
\usage{
reference_Poisson
}
\description{
Poisson regression reference coefficients for test without sample weights.
}
\keyword{datasets}
