% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{sampleWeights}
\alias{sampleWeights}
\title{Sample weights}
\format{Length 1000 numeric vector}
\usage{
sampleWeights
}
\description{
Length 1000 numeric vector of sample weights for the examples.
}
\keyword{datasets}
