% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sempl.R
\name{sempl}
\alias{sempl}
\title{Fitting Linear Structural Equation Models}
\usage{
sempl(Y, B, Omega, B.hat = NULL, Omega.hat = NULL, mu.hat = NULL,
  meanEst = T, covarRestrict = NULL, type = "profile", tol = 1e-04,
  maxIter = 500, outerIters = 200)
}
\arguments{
\item{Y}{p x n matrix with data}

\item{B}{V by V matrix with {0,1} giving structure of directed edges}

\item{Omega}{V by V matrix with {0,1} giving structure of bi-directed edges}

\item{B.hat}{V by V matrix giving initial edges weights for directed edges. If BInit is NULL,
a default initialization will be used.}

\item{Omega.hat}{V by V matrix giving initial edge weights for bi-directed edges}

\item{mu.hat}{Initial value for means}

\item{type}{profile, naive, euclid}
}
\description{
Fits Empirical Likelihood estimates for given SEM
}

