% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sempl_cd.R
\name{sempl_cd}
\alias{sempl_cd}
\title{Fitting Linear Structural Equation Models}
\usage{
sempl_cd(Y, B, Omega, covarRestrict, meanEst = F, tol = 0.001,
  outerTol = 0.01)
}
\arguments{
\item{Y}{p x n matrix with data}

\item{B}{V by V matrix with {0,1} giving structure of directed edges}
}
\description{
Fits Empirical Likelihood estimates for given SEM
}

