% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CI.R
\name{var.el}
\alias{var.el}
\title{Empirical Likelihood Methods for Linear Structural Equation Models}
\usage{
var.el(Y, B, Omega, B.hat, Omega.hat = NULL, p = NULL, cutoff = NULL,
  grid = 0.005, type = "asymptotic")
}
\arguments{
\item{Y}{V by N matrix with observed data}

\item{B}{V by V matrix with {0,1} giving structure of directed edges}

\item{Omega}{V by V matrix with {0,1} giving structure of bi-directed edges}

\item{B.hat}{V by V matrix giving estimated edges weights for directed edges}

\item{Omega.hat}{V by V matrix giving estimated edge weights for bi-directed edges}

\item{p}{vector of length N containing the estimated empirical likelihood weights}

\item{type}{the type of variances to return. options are "asymptotic" for the asymptotic
variance as defined in Qin and Lawless 1994 or "profile" for profile confiendence intervals}
}
\value{
The (scaled by n) estimated covariance matrix or profile confidence intervals
}
\description{
Calculate assymptotic variance for non-structural zeros in B and Omega matrix. The ordering of the estimates
in the variance matrix follows standard vec(-) convention indexing over rows first then columns. The elements of B
are ordered before the elements of Omega. Duplicate elements of Omega are not included.
}

