% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CI.R
\name{var.ricf}
\alias{var.ricf}
\title{Maximum Likelihood Method for Linear Structural Equation Models}
\usage{
var.ricf(Y, B, Omega, B.hat, Omega.hat, type = "expected")
}
\arguments{
\item{B}{V by V matrix with {0,1} giving structure of directed edges}

\item{Omega}{V by V matrix with {0,1} giving structure of bi-directed edges}

\item{B.hat}{V by V matrix giving estimated edges weights for directed edges}

\item{Omega.hat}{V by V matrix giving estimated edge weights for bi-directed edges}

\item{type}{string describing which Fisher Information to calculate. Options are "expected" or "observed".}

\item{S}{V by V matrix corresponding to the sample covariance}
}
\value{
The inverse (scaled by n) Fisher information matrix as derived by Fox and Drton 2014 or the Huber-White misspecified model covariance estimate
}
\description{
Calculate Fisher information for non-structural zeros in B and Omega matrix. The ordering of the estimates
in the Information matrix follows standard vec(-) convention indexing over rows first then columns. The elements of B
are ordered before the elements of Omega. Duplicate elements of Omega are not included.
}

