% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{X}
\alias{X}
\title{Design matrix for tests and examples}
\format{
A 1000 by 50 matrix
}
\usage{
X
}
\description{
A 1000 by 50 matrix with 1's for the intercept in the 
first column and numeric values in the remaining columns. 
Used in tests and examples
}
\keyword{datasets}
