% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{y_binary}
\alias{y_binary}
\title{Binary response vector}
\format{
Length 1000 integer vector
}
\usage{
y_binary
}
\description{
Length 1000 vector of 0 and 1 values for logistic 
regression example.
}
\keyword{datasets}
