% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{y_count}
\alias{y_count}
\title{Poisson response vector}
\format{
Length 1000 numeric vector
}
\usage{
y_count
}
\description{
Length 1000 integer vector containing responses for the Poisson regression example.
}
\keyword{datasets}
