% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{y_gaussian}
\alias{y_gaussian}
\title{Gaussian response vector}
\format{
Length 1000 numeric vector
}
\usage{
y_gaussian
}
\description{
Length 1000 numeric vector containing responses for the 
linear regression example.
}
\keyword{datasets}
