% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{y_multinomial}
\alias{y_multinomial}
\title{Multinomial response example}
\format{Length 1000 numeric vector}
\usage{
y_multinomial
}
\description{
Length 1000 vector with values 1, 2, and 3 for the multinomial
regression example.
}
\keyword{datasets}
