% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{y_multiresponse}
\alias{y_multiresponse}
\title{Multiresponse Gaussian response matrix}
\format{
1000 by 3 numeric matrix
}
\usage{
y_multiresponse
}
\description{
1000 by 3 numeric matrix containing responses for the multiresponse 
Gaussian example
}
\keyword{datasets}
