% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FTTest.R
\name{FTtest}
\alias{FTtest}
\title{Freeman–Tukey  Test for Bivariate Distributions via Conditional Specification}
\usage{
FTtest(data, distribution, params, num_params)
}
\arguments{
\item{data}{a dataset or matrix with two columns.}

\item{distribution}{a string specifying the theoretical distribution (`"BBCD"`, `"BBPD"`, or `"BBGD"`).}

\item{params}{a named list of parameters required by the specified distribution.}

\item{num_params}{an integer specifying the number of parameters that were estimated}
}
\value{
A list with components:
\describe{
  \item{observed}{Observed frequency table}
  \item{expected}{Expected frequency table under the specified distribution}
  \item{test}{Result of the Freeman–Tukey test, a list with test statistic and p-value}
}
}
\description{
Performs a goodness-of-fit test using the Freeman–Tukey (F–T) statistic
for a given dataset and a specified bivariate distribution.
}
\examples{
samples <- rbinomBCD(n = 50, n1 = 5, n2 = 2, p1 = 0.15, p2 = 0.9, lambda = 1, seed = 1)
params1 <- MLEbinomBCD(samples, fixed_n1 = 5, fixed_n2 = 2)
result_bbcd1 <- FTtest(samples, "BBCD", params1, num_params = 3)
result_bbcd1

params2 <- MLEbinomBCD(samples)
result_bbcd2 <- FTtest(samples, "BBCD", params2, num_params = 5)
result_bbcd2

samples <- rgeomBCD(n = 50, q1 = 0.5, q2 = 0.5, q3 = 0.1, seed = 123)
params <- MLEgeomBCD(samples)
result_bgcd <- FTtest(samples, "BGCD", params, num_params = 3)
result_bgcd

samples <- rpoisBCD(n = 50, lambda1 = 0.5, lambda2 = 0.5, lambda3 = 0.5, seed = 8)
params <- MLEpoisBCD(samples)
result_bpcd <- FTtest(samples, "BPCD", params, num_params = 3)
result_bpcd

}
