% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ricf.R
\name{ricf}
\alias{ricf}
\title{Fitting Cyclic Linear Structural Equation Models}
\usage{
ricf(B, Omega, Y, BInit = NULL, OmegaInit = NULL, sigConv = TRUE,
  tol = 1e-05, maxIter = 5000, out = "none", maxKap = 10^9)
}
\arguments{
\item{B}{V by V matrix with {0,1} giving structure of directed edges}

\item{Omega}{V by V matrix with {0,1} giving structure of bi-directed edges}

\item{BInit}{V by V matrix giving initial edges weights for directed edges}

\item{OmegaInit}{V by V matrix giving initial edge weights for bi-directed edges}

\item{tol}{convegence tolerance}

\item{maxIter}{integer specifying the maximum number of iterations}

\item{maxKap}{positive scalar which determines largest conditioning number before throwing error}

\item{signConv}{boolean which specifies how to measure convergence. \code{TRUE} looks for
convergence in Sigma while \code{FALSE} looks for convergence in the actual edge weight estimates}
}
\value{
\item{sigmaHat}{estimated covariance matrix at convergence}
   \item{bHat}{estimated B matrix (edge weights for directed edges) at converegence}
   \item{omegaHat}{estimated Omega (edge weights for bi-directed edges) at convergence}
   \item{iterations}{number of iterations until convergence. a single iteration is considered
   a pass through all nodes}
   \item{converged}{boolean on whether or not the algorithm converged before the max iterations}
}
\description{
Estimates MLE's for cyclic linear SEM's as described in DFW
}

